Add CUDA support for upscaling technology
This update introduces CUDA as a new upscaling technology option within the application. It includes the addition of CUDA-related dependencies in Cargo.toml and Cargo.lock, as well as modifications to the AppState and UI to allow users to select CUDA for GPU acceleration. The implementation also includes a fallback mechanism to other upscaling technologies if CUDA is not supported, enhancing the overall flexibility and performance of the upscaling process.